% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggStyle.R
\name{ggStyle}
\alias{ggStyle}
\title{Apply personal style options for ggplot2 objects}
\usage{
ggStyle(filename, sizefac = 1)
}
\arguments{
\item{filename}{Path to filename, will be displayed as caption. Character string.}

\item{sizefac}{Factor by which the element sizes are being multiplied with. Numeric.}
}
\value{
list of ggplot2 options
}
\description{
This function returns some common options to ggplot2 objects.
}
\examples{
if(FALSE){
 ggplot(iris, aes(x=Sepal.Length, y=Sepal.Width))+
 geom_point()+
 ggStyle(filename="path/Scripts/file.qmd", sizefac=1.2)
}
}
