% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderQuartoMove.R
\name{renderQuartoMove}
\alias{renderQuartoMove}
\title{Render .qmd file and move to target location}
\usage{
renderQuartoMove(input, output_file = NULL, output_dir = NULL, ...)
}
\arguments{
\item{input}{What is the filename called? Needs to contain .qmd at the end. Character string}

\item{output_file}{How should the output file be named? Needs to contain file suffix. Character string}

\item{output_dir}{Relative or absolute path to the desired output directory. Character string.}

\item{...}{further arguments passed to quarto::quarto_render()}
}
\value{
none
}
\description{
This is a helper function as long as quarto does not natively supports
moving files to a certain location after rendering.
see https://github.com/quarto-dev/quarto-cli/discussions/2171 and
}
\examples{
if (FALSE) {
 # save timestamp for output file
 timestamp <- Sys.time() \%>\% format("\%d-\%m-\%Y_\%H-\%M")

 # define output file
 output_file <- paste0("ModifiedPriors_FullBayes_" , timestamp,  ".html")

 # run quarto and move fileto target location
 quarto_render_move(input="ModifiedPriors_FullBayes.qmd",
              output_format="html",
              output_file = output_file,
              output_dir = "../Output/")
}
}
