% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sayHelloWorld.R
\name{sayHelloWorld}
\alias{sayHelloWorld}
\title{Say hello to the new world and choose to what world you want to say hello.}
\usage{
sayHelloWorld(name = "earth")
}
\arguments{
\item{name}{Name of the world for which a hello statement should
be written. Needs to be a character.}
}
\value{
Character string based on the name variable and a greeting formula.
}
\description{
Say hello to the new world and choose to what world you want to say hello.
}
\examples{
sayHelloWorld()
sayHelloWorld("jupiter")
sayHelloWorld(name="mars")
}
