% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sayNumber.R
\name{sayNumber}
\alias{sayNumber}
\title{Print out a certain number with text.}
\usage{
sayNumber(num = 1)
}
\arguments{
\item{num}{The number which should be displayed. Numeric.}
}
\value{
Text output containing text and the number.
}
\description{
Print out a certain number with text.
}
\examples{
sayNumber()
sayNumber(2)
sayNumber(num=923)
}
